requirements.txt
Flask==2.2.3
psycopg2-binary==2.9.9
Werkzeug==2.2.3